+++ python/cudf/cudf/core/dataframe.py
@@ -9,7 +9,6 @@ import sys
 import warnings
 from collections import OrderedDict, defaultdict
 from collections.abc import Mapping, Sequence
-from types import GeneratorType
 
 import cupy
 import numpy as np
@@ -205,7 +204,7 @@ class DataFrame(Frame, Serializable):
             self._data = data
             if index is None:
                 index = as_index(range(self._data.nrows))
-            self._index = as_index(index)
+            self.index = as_index(index)
             return None
 
         if isinstance(data, DataFrame):
@@ -255,7 +254,7 @@ class DataFrame(Frame, Serializable):
                 new_df = self._from_arrays(data, index=index, columns=columns)
 
             self._data = new_df._data
-            self._index = new_df._index
+            self.index = new_df._index
             self.columns = new_df.columns
         elif hasattr(data, "__array_interface__"):
             arr_interface = data.__array_interface__
@@ -265,7 +264,7 @@ class DataFrame(Frame, Serializable):
             else:
                 new_df = self.from_records(data, index=index, columns=columns)
             self._data = new_df._data
-            self._index = new_df._index
+            self.index = new_df._index
             self.columns = new_df.columns
         else:
             if is_list_like(data):
@@ -274,7 +273,7 @@ class DataFrame(Frame, Serializable):
                         [data], index=index, columns=columns
                     )
                     self._data = new_df._data
-                    self._index = new_df._index
+                    self.index = new_df._index
                     self.columns = new_df.columns
                 elif len(data) > 0 and isinstance(data[0], Series):
                     self._init_from_series_list(
@@ -2475,10 +2474,13 @@ class DataFrame(Frame, Serializable):
 
     @index.setter
     def index(self, value):
+        old_length = (
+            self._num_rows if self._index is None else len(self._index)
+        )
         if isinstance(value, cudf.core.multiindex.MultiIndex):
-            if len(self._data) > 0 and len(value) != len(self):
+            if len(self._data) > 0 and len(value) != old_length:
                 msg = (
-                    f"Length mismatch: Expected axis has {len(self)} "
+                    f"Length mismatch: Expected axis has {old_length} "
                     f"elements, new values have {len(value)} elements"
                 )
                 raise ValueError(msg)
@@ -2486,7 +2488,6 @@ class DataFrame(Frame, Serializable):
             return
 
         new_length = len(value)
-        old_length = len(self._index)
 
         if len(self._data) > 0 and new_length != old_length:
             msg = (
@@ -2626,7 +2627,7 @@ class DataFrame(Frame, Serializable):
 
     def set_index(
         self,
-        index,
+        keys,
         drop=True,
         append=False,
         inplace=False,
@@ -2636,7 +2637,7 @@ class DataFrame(Frame, Serializable):
 
         Parameters
         ----------
-        index : Index, Series-convertible, label-like, or list
+        keys : Index, Series-convertible, label-like, or list
             Index : the new index.
             Series-convertible : values for the new index.
             Label-like : Label of column to be used as index.
@@ -2720,15 +2721,15 @@ class DataFrame(Frame, Serializable):
         5  e  5.0
         """
 
-        if not isinstance(index, list):
-            index = [index]
+        if not isinstance(keys, list):
+            keys = [keys]
 
         # Preliminary type check
         col_not_found = []
         columns_to_add = []
         names = []
         to_drop = []
-        for i, col in enumerate(index):
+        for i, col in enumerate(keys):
             # Is column label
             if is_scalar(col) or isinstance(col, tuple):
                 if col in self.columns:
@@ -3003,30 +3004,6 @@ class DataFrame(Frame, Serializable):
         value = column.as_column(value)
 
         self._data.insert(name, value, loc=loc)
-
-    def add_column(self, name, data, forceindex=False):
-        """Add a column
-
-        Parameters
-        ----------
-        name : str
-            Name of column to be added.
-        data : Series, array-like
-            Values to be added.
-        """
-
-        warnings.warn(
-            "`add_column` will be removed in the future. Use `.insert`",
-            DeprecationWarning,
-        )
-
-        if name in self._data:
-            raise NameError("duplicated column name {!r}".format(name))
-
-        if isinstance(data, GeneratorType):
-            data = Series(data)
-
-        self.insert(len(self._data.names), name, data)
 
     def drop(
         self,
@@ -3869,7 +3846,6 @@ class DataFrame(Frame, Serializable):
         sort=False,
         lsuffix=None,
         rsuffix=None,
-        type="",
         method="hash",
         indicator=False,
         suffixes=("_x", "_y"),
@@ -3960,14 +3936,6 @@ class DataFrame(Frame, Serializable):
         else:
             lsuffix, rsuffix = suffixes
 
-        if type != "":
-            warnings.warn(
-                'type="' + type + '" parameter is deprecated.'
-                'Use method="' + type + '" instead.',
-                DeprecationWarning,
-            )
-            method = type
-
         lhs = self.copy(deep=False)
         rhs = right.copy(deep=False)
 
@@ -3998,7 +3966,6 @@ class DataFrame(Frame, Serializable):
         lsuffix="",
         rsuffix="",
         sort=False,
-        type="",
         method="hash",
     ):
         """Join columns with other DataFrame on index or on a key column.
@@ -4025,14 +3992,6 @@ class DataFrame(Frame, Serializable):
         - *other* must be a single DataFrame for now.
         - *on* is not supported yet due to lack of multi-index support.
         """
-        # Outer joins still use the old implementation
-        if type != "":
-            warnings.warn(
-                'type="' + type + '" parameter is deprecated.'
-                'Use method="' + type + '" instead.',
-                DeprecationWarning,
-            )
-            method = type
 
         lhs = self
         rhs = other
@@ -4063,7 +4022,6 @@ class DataFrame(Frame, Serializable):
         squeeze=False,
         observed=False,
         dropna=True,
-        method=None,
     ):
         if axis not in (0, "index"):
             raise NotImplementedError("axis parameter is not yet implemented")
@@ -4087,11 +4045,6 @@ class DataFrame(Frame, Serializable):
                 "groupby() requires either by or level to be" "specified."
             )
 
-        if method is not None:
-            warnings.warn(
-                "The 'method' argument is deprecated and will be unused",
-                DeprecationWarning,
-            )
         return DataFrameGroupBy(
             self,
             by=by,
@@ -5215,87 +5168,6 @@ class DataFrame(Frame, Serializable):
         else:
             df._index = as_index(index)
         return df
-
-    @classmethod
-    def from_gpu_matrix(
-        self, data, index=None, columns=None, nan_as_null=False
-    ):
-        """Convert from a numba gpu ndarray.
-
-        Parameters
-        ----------
-        data : numba gpu ndarray
-        index : str, Index
-            The name of the index column in `data` or an Index itself.
-            If None, the default index is used.
-        columns : list of str
-            List of column names to include.
-
-        Returns
-        -------
-        DataFrame
-        """
-        warnings.warn(
-            "DataFrame.from_gpu_matrix will be removed in 0.16. "
-            "Please use cudf.DataFrame() to create a DataFrame "
-            "out of a gpu matrix",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-        if data.ndim != 2:
-            raise ValueError(
-                f"matrix dimension expected 2 but found {data.ndim}"
-            )
-
-        if columns is None:
-            names = [i for i in range(data.shape[1])]
-        else:
-            if len(columns) != data.shape[1]:
-                raise ValueError(
-                    f"columns length expected {data.shape[1]} but "
-                    f"found {len(columns)}"
-                )
-            names = columns
-
-        if (
-            index is not None
-            and not isinstance(index, (str, int))
-            and len(index) != data.shape[0]
-        ):
-            raise ValueError(
-                f"index length expected {data.shape[0]} but found {len(index)}"
-            )
-
-        df = DataFrame()
-        data = cupy.asfortranarray(cupy.asarray(data))
-        for i, k in enumerate(names):
-            df._data[k] = as_column(data[:, i], nan_as_null=nan_as_null)
-
-        if index is not None:
-            if isinstance(index, (str, int)):
-                index = as_index(df[index])
-            else:
-                index = as_index(index)
-        else:
-            index = RangeIndex(start=0, stop=len(data))
-        df._index = index
-
-        return df
-
-    def to_gpu_matrix(self):
-        """Convert to a numba gpu ndarray
-
-        Returns
-        -------
-        numba gpu ndarray
-        """
-        warnings.warn(
-            "The to_gpu_matrix method will be deprecated"
-            "in the future. use as_gpu_matrix instead.",
-            DeprecationWarning,
-        )
-        return self.as_gpu_matrix()
 
     @classmethod
     def _from_columns(cls, cols, index=None, columns=None):
@@ -6554,9 +6426,7 @@ class DataFrame(Frame, Serializable):
                     )
                 return Series(result, index=self.index, dtype=result_dtype,)
             else:
-                result_df = DataFrame.from_gpu_matrix(result).set_index(
-                    self.index
-                )
+                result_df = DataFrame(result).set_index(self.index)
                 result_df.columns = prepared.columns
                 return result_df
 
@@ -6841,9 +6711,7 @@ class DataFrame(Frame, Serializable):
         cov : DataFrame
         """
         cov = cupy.cov(self.values, rowvar=False)
-        df = DataFrame.from_gpu_matrix(cupy.asfortranarray(cov)).set_index(
-            self.columns
-        )
+        df = DataFrame(cupy.asfortranarray(cov)).set_index(self.columns)
         df.columns = self.columns
         return df
 
@@ -6851,9 +6719,7 @@ class DataFrame(Frame, Serializable):
         """Compute the correlation matrix of a DataFrame.
         """
         corr = cupy.corrcoef(self.values, rowvar=False)
-        df = DataFrame.from_gpu_matrix(cupy.asfortranarray(corr)).set_index(
-            self.columns
-        )
+        df = DataFrame(cupy.asfortranarray(corr)).set_index(self.columns)
         df.columns = self.columns
         return df